+++ web/src/client/css/layout/_forms.scss
@@ -67,15 +67,6 @@ form {
       }
     }
 
-    &__errors {
-      color: $danger;
-      ul {
-        list-style: none;
-        margin: 0;
-        padding: 0;
-      }
-    }
-
   }
 
   label {
