+++ web/src/client/js/components/Form/TextField.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import cx from 'classnames'
-import FieldErrors from './FieldErrors'
+import ValidationComponent from 'Components/Validation/ValidationComponent'
 
 const TextField = ({ help, id, label, large, name, errors, placeholder, value, onChange }) => {
   const textFieldClasses = cx({
@@ -21,7 +21,7 @@ const TextField = ({ help, id, label, large, name, errors, placeholder, value, o
           placeholder={placeholder}
           defaultValue={value}
           onChange={onChange} />
-        <FieldErrors errors={errors} />
+        <ValidationComponent errors={errors} />
       </div>
       {help && <div className="form-field__help small">{help}</div>}
     </div>
