+++ web/src/client/js/components/Validation/Validation.scss
@@ -1,0 +1,12 @@
+@import 'CSS/inc/_vars.scss';
+
+.validation {
+  &__errors {
+    color: $danger;
+    ul {
+      list-style: none;
+      margin: 0;
+      padding: 0;
+    }
+  }
+}
